Move Firefox profile from ceno-box to browser-profiles; Improvements in CENOBox build and client scripts:
CENOBox/CENO.sh
  * Load CENO Router extension on the new Chrome window/profile
  * Fix: Exporting CENOLANG to agents
  * Improved control over starting/killing CENOClient

ceno-box/build.sh:
  * Stop CENOBox if running, before removing the directory
  * Make directory for browser profiles in CENOBox

+++ ceno-box/ceno-extra/CENO.sh
@@ -23,8 +23,7 @@ function getEnvLang {
       CENOLANG=en-us
     fi
   fi
-  echo "CENO language set to" $CENOLANG
-  export CENOLANG
+  echo $CENOLANG
 }
 
 function browserExists {
@@ -37,26 +36,33 @@ function browserExists {
 }
 
 function startChromeProfile {
-  $1 --profile-directory=ceno-chrome --incognito ./ceno-client/views/extension-en-us.html &> /dev/null &
+  #Experimental: Load the CENO Router extension while invoking chrome
+  tempProfile=$(mktemp -d browser-profiles/chrome/google-chrome.XXXXXXX)
+  $1 --profile-directory=$tempProfile --load-extension=$(pwd)/browser-extensions/ceno-chrome --no-first-run &> /dev/null &
+  rm $tempProfile
 }
 
 function startBrowser {
+  extIntaller=./ceno-client/views/extension-en-us.html
   # Open a browser window with the CENO profiles, including the plugin
   if browserExists chrome
   then
-    startChromeProfile chrome
+    startChromeProfile chrome $extIntaller
   elif browserExists chromium-browser
   then
-    startChromeProfile chromium-browser
+    startChromeProfile chromium-browser $extIntaller
   elif browserExists chromium
   then
-    startChromeProfile chromium
+    startChromeProfile chromium $extIntaller
+  elif browserExists Chromium
+  then
+    startChromeProfile Chromium $extIntaller
   elif browserExists google-chrome
   then
-    startChromeProfile google-chrome
+    startChromeProfile google-chrome $extIntaller
   elif browserExists firefox
   then
-      firefox -no-remote -private-window -profile "ceno-firefox" ./ceno-client/views/extension-en-us.html &> /dev/null &
+      firefox -no-remote -private-window -profile "browser-profiles/firefox" $extIntaller &> /dev/null &
   else
       echo "None of the supported browsers is installed in your machine."
       echo "Please install Chrome or Firefox and execute this script again."
@@ -69,13 +75,18 @@ function startCENO {
   ./run.sh start &> CENO.log
 
   # Start CENOClient proxy
-  cd ceno-client
-  CENOLANG=en-us ./CENOClient &> ../CENO.log &
-  CENOClientPID=$!
-  echo $CENOClientPID > CENOClient.pid
-  kill -20 $CENOClientPID
-  kill -18 $CENOClientPID
-  cd ..
+  if ps ax | grep -v grep | grep CENOClient > /dev/null
+  then
+    echo "CENOClient is already running"
+  else
+    cd ceno-client
+    CENOLANG=en-us ./CENOClient &> ../CENO.log &
+    CENOClientPID=$!
+    echo $CENOClientPID > CENOClient.pid
+    kill -18 $CENOClientPID
+    cd ..
+    echo "Started CENOClient proxy"
+  fi
 
   echo "You are ready to use CENO."
   echo "Remember that you are covered by CENO only when the CENO Router plugin is loaded in your browser."
@@ -85,11 +96,17 @@ function startCENO {
 case "$1" in
   'stop')
     ./run.sh stop
-    kill $(cat ceno-client/CENOClient.pid)
+    if [ -d ceno-client/CENOClient.pid ]
+    then
+      kill $(cat ceno-client/CENOClient.pid)
+      echo "Stopped CENO Client proxy"
+    fi
     ;;
 
   *)
-    getEnvLang
+    CENOLANG=$(getEnvLang)
+    export CENOLANG
+    echo "CENO language set to" $CENOLANG
     startBrowser
     startCENO
     ;;
